+++ tests/fuzz/fuzz.c
@@ -444,7 +444,31 @@ static int
     return 0;
 }
 
+static int
+    fuzz_decode_pldm_rde_rde_operation_complete_req(const struct pldm_msg* msg,
+                                                    size_t payload_length)
+{
+    struct pldm_rde_rde_operation_complete_req req;
+
+    decode_pldm_rde_rde_operation_complete_req(msg, payload_length, &req);
+
+    return 0;
+}
+
+static int
+    fuzz_decode_pldm_rde_rde_operation_complete_resp(const struct pldm_msg* msg,
+                                                     size_t payload_length)
+{
+    struct pldm_rde_rde_operation_complete_resp resp;
+
+    decode_pldm_rde_rde_operation_complete_resp(msg, payload_length, &resp);
+
+    return 0;
+}
+
 static int (*const decode_pldm_msg_tests[])(const struct pldm_msg*, size_t) = {
+    fuzz_decode_pldm_rde_rde_operation_complete_req,
+    fuzz_decode_pldm_rde_rde_operation_complete_resp,
     fuzz_decode_pldm_rde_rde_operation_init_req,
     fuzz_decode_pldm_rde_rde_operation_init_resp,
     fuzz_decode_pldm_rde_get_resource_etag_req,